+++ Cargo.toml
@@ -157,6 +157,8 @@ uuid = { version = "0.8.1", features = ["serde", "v4"] }
 vlog = "0.1.4"
 walkdir = "2"
 minitrace = { git = "https://github.com/pingcap-incubator/minitrace-rust.git", branch = "master" }
+flog = "0.1.1"
+minstant = { git = "https://github.com/zhongzc/minstant.git" }
 
 
 [dependencies.prometheus-static-metric]
